% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.data.frame.bsmdata}
\alias{as.data.frame.bsmdata}
\title{Convert to Data Frame}
\usage{
\method{as.data.frame}{bsmdata}(x, ...)
}
\arguments{
\item{x}{bsmdata object}

\item{...}{extra arguments}
}
\description{
Convert a bsmdata object into a dataframe
}
\author{
Tom Elliott
}

