% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.mcmc.bsmfit}
\alias{as.mcmc.bsmfit}
\title{Convert BSMFit to MCMC}
\usage{
\method{as.mcmc}{bsmfit}(x)
}
\arguments{
\item{x}{bsmfit object}
}
\value{
MCMC list
}
\description{
Convert a bsmfit object into an MCMC list for analysis using the CODA package
}
\details{
Simply converts the \code{bsmfit} object from \code{bsm} into an \code{mcmc.list} object. The
result can then be used with the \code{CODA} package, specifically the function
\code{codamenu()}.
}
\author{
Tom Elliott
}

