% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootsvmmrmrwt.R
\name{bootsvmmrmrwt}
\alias{bootsvmmrmrwt}
\title{Computation weights for gene selection through Bootstrap (Boot), Support Vector Machine (SVM)  and Maximum Relevance and Minimum Redundancy (MRMR) methods.}
\usage{
bootsvmmrmrwt(x, y, method, beta, nboot)
}
\arguments{
\item{x}{Nxp data frame of gene expression values, where, N represents number of genes and p represents samples/time points generated in a case vs. control gene expression study.}

\item{y}{px1 numeric vector with entries 1 and -1 representing sample/subject labels, where 1 and -1 represents the labels of subjects/ samples for case and control conditions respectively.}

\item{method}{Character variable representing either 'Linear' or 'Quadratic' method for integrating the weights/scores computed through SVM and MRMR methods.}

\item{beta}{scalar representing trade-off between SVM and MRMR weights.}

\item{nboot}{scalar representing the number of bootsrap samples to be drawn from the data using simple random sampling with replacement (Bootstrap) procedure.}
}
\value{
This returns a vector of bootstrap weights for gene selection computed through BSM approach.
}
\description{
The function computes the integrated bootstrap weights for gene selection using the Bootstrap-Support Vector Machine (SVM)-Maximum Relevance and Minimum Redundancy (MRMR) (BSM) approach from high dimensional gene expression data .
}
\details{
Computation weights for gene selection through BSM approach.
Takes the gene expression data matrix (rows as genes and coloumns as samples) and vector of class labels of subjects (1: case and -1: control) as inputs.
}
\examples{
x=as.data.frame(matrix(runif(1000), 50))
row.names(x) = paste("Gene", 1:50)
colnames(x) = paste("Samp", 1:20)
y=as.numeric(c(rep(1, 10), rep(-1, 10)))
bootsvmmrmrwt(x, y, method="Linear", beta=0.6, nboot=20)

}
\author{
Samarendra Das <samarendra4849 at gamil.com>
}
