% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{BSM-Package}
\alias{BSM-Package}
\alias{BSM-Package-package}
\title{Bayesian Selectivity Modelling of Fisheries Trawl Data}
\description{
This package allows users to easily fit Bayesian models to fisheries trawl data. It implements
its own data structure, plot methods, and various other methods to make it as user-friendly as
possible by using functions users know (eg. \code{plot}).
}
\author{
Tom Elliott
}
\references{
My MSc thesis.
}

