% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bsmCurve}
\alias{bsmCurve}
\title{Draw Selection Curves}
\usage{
bsmCurve(x, obj, xlim, add = TRUE, ...)
}
\arguments{
\item{x}{vector of L50, SR, and delta (optional)}

\item{obj}{a \code{bsmdata} object that includes the data, which will be drawn if \code{add = FALSE}}

\item{xlim}{numeric vector of length 2, specifying the range of x values to draw the curve over}

\item{add}{logical, if \code{TRUE}, the curve is added to an existing plot, otherwise a new one
is drawn.}

\item{...}{additional arguments}
}
\description{
Draws a selection curve for a given set of parameter values.
}
\details{
If \code{obj} if specified, then the appropriate axes are drawn using that. Alternatively,
generic x-limits can be specified by \code{xlim}. If these are both empty, but \code{add =
TRUE}, then the values are taken from \code{par()}.
}
\author{
Tom Elliott
}

