% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bsmData}
\alias{bsmData}
\title{Import Data for BSM}
\usage{
bsmData(y1, y2, n, length, length.unit = NULL, haul, q1, q2, paired = FALSE,
  data = NULL, ...)
}
\arguments{
\item{y1}{the count for the codend or experimental trawl}

\item{y2}{the count fo the cover or control trawl}

\item{n}{the total fish caught in the trawl}

\item{length}{the length of fish}

\item{length.unit}{the unit of length measurements}

\item{haul}{the haul id}

\item{q1}{sampling fraction for codend/experimental trawl}

\item{q2}{sampling fraction for cover/experimental trawl}

\item{paired}{logical, is the data from a paired trawl?}

\item{data}{the name of the data frame. If \code{NULL}, then the other supplied argument
must be vectors. Otherwise if \code{df} is supplied, then the other arguments are the
variable names.}

\item{...}{additional variables can be passed to the object, such as weights, mesh sizes, and
any other covariates to be used in later modelling}
}
\value{
An object of class \code{bsmdf}. This is a list with necessary information,
and is what should be supplied as the \code{data} argument to other
\code{\link{bsmData}} functions.
}
\description{
This function handles manipulation of a provided data set (in the form of a data
frame, from \code{read.csv} or similar) to a form that can be used by other
\code{\link{bsmData}} functions.
}
\details{
The simplest use is where a data frame is provided, and the required column names are
specified as function arguments.

It is also possible to supply vectors and no data frame.
}
\author{
Tom Elliott
}

