% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dic.bsmfit}
\alias{dic.bsmfit}
\title{Get DIC Value}
\usage{
\method{dic}{bsmfit}(x, ...)
}
\arguments{
\item{x}{a bsm.fit object}

\item{...}{extra arguments}
}
\value{
list
}
\description{
Extract the DIC value from the bsm.fit object
}
\details{
Does what it says
}
\author{
tell029
}

