% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dic}
\alias{dic}
\alias{dic.bsmfit}
\title{Extract DIC Value}
\usage{
dic(x, ...)

\method{dic}{bsmfit}(x, ...)
}
\arguments{
\item{x}{an object, such as bsmfit}

\item{...}{extra arguments}
}
\value{
the DIC summary information
}
\description{
Extract the DIC value from an object
}
\details{
Generic Method
}
\section{Methods (by class)}{
\itemize{
\item \code{bsmfit}: Extract the DIC information from the JAGS object
}}
\author{
Tom Elliott
}

