% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generateFormula}
\alias{generateFormula}
\alias{print.bsmFormula}
\title{Generate a mathematical expression for a parameter}
\usage{
generateFormula(par, expression, data, est, ...)

\method{print}{bsmFormula}(x, use.values = FALSE, ...)
}
\arguments{
\item{par}{parameter being formulated}

\item{expression}{an expression}

\item{data}{data from fitted object}

\item{est}{parameter estimates for formula}

\item{...}{additional arguments}

\item{x}{object of class bsmFormula}

\item{use.values}{if \code{TRUE}, the parameter estimates are used}
}
\value{
a bsmFormula object for a given parameter
}
\description{
.. content for description{} (no empty lines) ..
}
\details{
.. content for details{} ..
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: 
}}
\author{
Tom Elliott
}

