% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getMean}
\alias{getMean}
\title{Get the Mean of Covariates}
\usage{
getMean(x, par, ...)
}
\arguments{
\item{x}{bsmfit object}

\item{par}{the parameter for means}

\item{...}{extra args}
}
\value{
numeric, the mean of the covariates
}
\description{
description of thsi
}
\details{
details of this
}
\author{
Tom Elliott
}

