% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getPars}
\alias{getPars}
\title{Parameter Names for the Model}
\usage{
getPars(family = "binomial", curve = "logistic", check.od = FALSE,
  od = FALSE, combine = is.null(random), random = NULL, L50 = NULL,
  SR = NULL, phi = NULL, delta = NULL, length.dist = "iid",
  paired = FALSE, ...)
}
\arguments{
\item{family}{\code{"logistic"} (default) or \code{"poisson"}, the family for the likelihood}

\item{curve}{\code{"logistic"} or \code{"richards"}, the type of selection curve to be fitted
(will likely allow other options, for example "Bspline", future)}

\item{check.od}{logical, if \code{TRUE}, then overdispersion estimates will be produced}

\item{od}{logical, if \code{TRUE}, the model will be fit allowing for overdispersion}

\item{combine}{logical, if \code{TRUE}, then the "combined hauls" approach is used, otherwise a
hierarchical approach is used.}

\item{random}{vector of parameters to have hierarchical or random effects}

\item{L50}{the formula for L50, default is L50 = ~haul}

\item{SR}{the formula for SR, default is SR = ~haul}

\item{phi}{the formula for phi, default is phi = ~haul}

\item{delta}{the formula for delta, default is delta = ~1}

\item{length.dist}{"iid" or "multinomial", the type of length distribution for the lambda
parameters. See details for more information.}

\item{paired}{whether or not the data are paired}

\item{...}{additional arguments}
}
\value{
lists of parameters
}
\description{
Returns lists of parameter names for summaries, initial values, etc.
}
\details{
Not much to know.
}
\author{
Tom Elliott
}

