% Generated by roxygen2 (4.0.1): do not edit by hand
\name{head.bsmdata}
\alias{head.bsmdata}
\title{Head of bsmdata.}
\usage{
\method{head}{bsmdata}(x, ...)
}
\arguments{
\item{x}{bsm object}

\item{...}{extra arguments}
}
\description{
Prints the head of a bsmdata object.
}
\author{
Tom Elliott
}

