% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeData}
\alias{makeData}
\title{Make JAGS Data from bsmdata Object}
\usage{
makeData(x, family = "binomial", combine = is.null(random), random = NULL,
  L50 = NULL, SR = NULL, phi = NULL, delta = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{bsmdata}, from the \code{bsmData} function}

\item{family}{\code{"logistic"} (default) or \code{"poisson"}, the family for the likelihood}

\item{combine}{logical, if \code{TRUE}, then the "combined hauls" approach is used, otherwise a
hierarchical approach is used.}

\item{random}{vector of parameters to have hierarchical or random effects}

\item{L50}{the formula for L50, default is L50 = ~haul if random = TRUE}

\item{SR}{the formula for SR, default is SR = ~haul}

\item{phi}{the formula for phi, default is phi = ~haul}

\item{delta}{the formula for delta, default is delta = ~1}

\item{...}{additional arguments}
}
\value{
A list of data for JAGS
}
\description{
Creates the required data list for JAGS for a given model.
}
\details{
Not much to know ...
}
\author{
Tom Elliott
}

