% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeDesign}
\alias{makeDesign}
\title{Make Design Matrix}
\usage{
makeDesign(par, dat)
}
\arguments{
\item{par}{the formula for the parameter}

\item{dat}{the dataframe}
}
\value{
design matrix
}
\description{
Makes a design matrix for given parameter.
}
\details{
Not much to see here ...
}
\author{
tell029
}

