% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeModel}
\alias{makeModel}
\title{Make JAGS Model}
\usage{
makeModel(family = "binomial", curve = "logistic", check.od = FALSE,
  od = FALSE, combine = is.null(random), random = NULL, L50 = NULL,
  SR = NULL, phi = NULL, delta = NULL, priors = NULL,
  length.dist = "iid", paired, file = tempfile(), show.info = TRUE, ...)
}
\arguments{
\item{family}{\code{"logistic"} (default) or \code{"poisson"}, the family for the likelihood}

\item{curve}{\code{"logistic"} or \code{"richards"}, the type of selection curve to be fitted
(will likely allow other options, for example "Bspline", future)}

\item{check.od}{logical, if \code{TRUE}, then overdispersion estimates will be produced}

\item{od}{logical, if \code{TRUE}, the model will be fit allowing for overdispersion}

\item{combine}{logical, if \code{TRUE}, then the "combined hauls" approach is used, otherwise a
hierarchical approach is used.}

\item{random}{vector of parameters to have hierarchical or random effects}

\item{L50}{the formula for L50, default is L50 = ~haul}

\item{SR}{the formula for SR, default is SR = ~haul}

\item{phi}{the formula for phi, default is phi = ~haul}

\item{delta}{the formula for delta, default is delta = ~1}

\item{priors}{the prior distributions for specified parameters. See details.}

\item{length.dist}{"iid" or "multinomial", the type of length distribution for the lambda
parameters. See details for more information.}

\item{paired}{whether the data are paired or not}

\item{file}{the file name to save the JAGS model in. If \code{NULL}, then a temporary file is
created}

\item{show.info}{logical, if \code{FALSE}, the package information will not be printed}

\item{...}{additional arguments}
}
\value{
NULL?
}
\description{
Write the model file in JAGS syntax.
}
\details{
A whole lot of fancyness that makes the JAGS model.
}
\author{
Tom Elliott
}

