% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.bsmdata}
\alias{plot.bsmdata}
\title{Plot Method for bsmdata Object}
\usage{
\method{plot}{bsmdata}(x, scale = TRUE, col, pch, legend = FALSE, weight,
  leg.posx = "topleft", leg.posy = NULL, leg.cex = 0.7, leg.bty = "n",
  ...)
}
\arguments{
\item{x}{a \code{bsmdata} object}

\item{scale}{logical, if \code{TRUE}, the data is scaled by sampling fractions}

\item{col}{the color of points. This can also be used to colour points by a factor variable such
as \code{haul}. See details.}

\item{pch}{the plotting symbol. This can also be the name of a (factor) variable in the data set}

\item{legend}{logical, if \code{TRUE} then a legend is drawn if points are coloured by some
variable (eg. haul)}

\item{weight}{the weights that will decide bubble sizes. Can be a variable or a vector}

\item{leg.posx}{position of legend. This can be text such as "topleft", "bottomright" etc, or
specific \code{x}-coordinates}

\item{leg.posy}{if \code{leg.posx} is an x-value, then this is the y-value}

\item{leg.cex}{size of legend}

\item{leg.bty}{the box type for the legend. Defaults to \code{"n"} for no box}

\item{...}{additional arguments that will be passed to the \code{plot.default} function}
}
\description{
A plot method to easily obtain a graphical representation of the data, with several options to
enable coding of some variables. For example, hauls can be colour coded, and weighting variables
used to control point size.
}
\details{
A lot of the models used are hierarchical, with random effects on hauls. Therefore, to see the
variation in hauls graphically, you can pass \code{col = "haul"} and the proportions will be
coloured differently for each haul.

The same idea is applied to plotting character (\code{pch}). To use a different plotting
character for each covariate (e.g., mesh size), simply pass \code{pch = mesh} into the
function. However, note that using \code{pch} and \code{col} will not be easy to see trends
unless they are very clear. It is recommended that you use both symbol and colour only if there
are individual points you wish to investigate, rather than for finding trends, which is best
done using multiple plots and colour.
}
\author{
Tom Elliott
}

