% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.bsmfit}
\alias{plot.bsmfit}
\title{Plot a BSM Fit}
\usage{
\method{plot}{bsmfit}(x, which = "posterior", parameters = NULL,
  estimate = "mean", cred.int = FALSE, cred.alpha = 0.95, new = TRUE,
  col = NULL, lty = 1, lwd = 2, legend.order = NULL,
  leg.posx = "topleft", leg.posy = NULL, leg.cex = 0.7, leg.bty = "n",
  chain.cols = rainbow(x$fit$BUGSoutput$n.chains + 1, s = 0.8, v = 0.8),
  interactive = FALSE, ...)
}
\arguments{
\item{x}{a bsmfit object}

\item{which}{which plot to draw, c("posterior", "curve")}

\item{parameters}{if \code{which = "posterior"}, \code{"pairs"}, then these parameters are used}

\item{estimate}{"mean" or "median", the estimate to be used when drawing the curve}

\item{cred.int}{logical, if \code{TRUE}, then a credible interval is calcualted from the
posterior samples of L50 and SR (and delta)}

\item{cred.alpha}{the significance level for the credible interval, default is 0.95}

\item{new}{logical, if \code{TRUE}, then a new plot of the data is drawn, otherwise the curve is
plot over any existing one}

\item{col}{colour of lines}

\item{lty}{line type}

\item{lwd}{line width}

\item{legend.order}{order of variables for legend purposes (color, line type, line width)}

\item{leg.posx}{x-position of legend, can be "topleft" etc}

\item{leg.posy}{NULL, or y-position of legend}

\item{leg.cex}{size of legend}

\item{leg.bty}{box type of legend}

\item{chain.cols}{colours of lines for traceplots and density plots, NOTE: not used for pairs
and curve}

\item{interactive}{logical, if \code{TRUE}, use clicking on the plot to navigate through plots}

\item{...}{additional parameters to \code{plot}}
}
\description{
Various types of plots of the fitted BSM object
}
\details{
Lots ??
}
\author{
Tom Elliott
}

