% Generated by roxygen2 (4.0.2): do not edit by hand
\name{postMean}
\alias{postMean}
\title{Get the posterior mean of a parameter}
\usage{
postMean(fit, par)
}
\arguments{
\item{fit}{a bsmfit object}

\item{par}{the parameter(s) to obtain means of}
}
\value{
numeric vector
}
\description{
Returns the BUGS posterior mean for a chosen parameter(s)
}
\details{
Just makes life easy ...
}
\author{
Tom Elliott
}

