% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predict.bsmfit}
\alias{predict.bsmfit}
\title{Predict parameter values for selection curve}
\usage{
\method{predict}{bsmfit}(object, sort = NULL, ...)
}
\arguments{
\item{object}{bsmfit}

\item{sort}{character vector of variables to sort by}

\item{...}{extra arguments}
}
\value{
matrix
}
\description{
.. content for description{} (no empty lines) ..
}
\details{
.. content for details{} ..
}
\author{
Tom Elliott
}

