% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predictPar}
\alias{predictPar}
\title{Predict the values for fixed factor levels}
\usage{
predictPar(x, ...)
}
\arguments{
\item{x}{bsmFormula object}

\item{...}{extra arguments}
}
\value{
Estimates
}
\description{
.. content for description{} (no empty lines) ..
}
\details{
.. content for details{} ..
}
\author{
Tom Elliott
}

