% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.bsm.predict}
\alias{print.bsm.predict}
\title{Print predicted parameter values for selection curve}
\usage{
\method{print}{bsm.predict}(x, ...)
}
\arguments{
\item{x}{bsm.predict object}

\item{...}{extra arguments}
}
\description{
.. content for description{} (no empty lines) ..
}
\details{
.. content for details{} ..
}
\author{
Tom Elliott
}

