% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.bsmdata}
\alias{print.bsmdata}
\title{Print bsmData Objectss}
\usage{
\method{print}{bsmdata}(x, ...)
}
\arguments{
\item{x}{a bsmdata object}

\item{...}{additional arguments (ignored)}
}
\description{
A print method for \code{bsmdata} objects.
}
\details{
Prints objects resulting from a call to \code{bsmData}.
}
\author{
Tom Elliott
}

