% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.bsmfit}
\alias{print.bsmfit}
\title{Print a bsmfit object}
\usage{
\method{print}{bsmfit}(x, model = FALSE, coda = FALSE, ...)
}
\arguments{
\item{x}{the result of a call to \code{bsm}}

\item{model}{logical, if \code{TRUE}, then the JAGS model is printed, otherwise summary
information is provided}

\item{coda}{logical, if \code{TRUE}, then the summary from the \code{coda} package is used}

\item{...}{additional parameters to \code{print.rjags}}
}
\description{
Print method of \code{bsmfit} object.
}
\details{
Not much yet...
}
\author{
Tom Elliott
}

