% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.bsmFormula}
\alias{print.bsmFormula}
\title{Print a BSM Formula}
\usage{
\method{print}{bsmFormula}(x, use.values = FALSE, ...)
}
\arguments{
\item{x}{object of class bsmFormula}

\item{use.values}{if \code{TRUE}, the parameter estimates are used}

\item{...}{extra args}
}
\description{
.. content for description{} (no empty lines) ..
}
\details{
.. content for details{} ..
}
\author{
Tom Elliott
}

