% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.dic.summary}
\alias{print.dic.summary}
\title{Print DIC output}
\usage{
\method{print}{dic.summary}(x, ...)
}
\arguments{
\item{x}{a dic.summary object}

\item{...}{extra arguments}
}
\description{
Print DIC information
}
\details{
Prints the results of dic()
}
\author{
tell029
}

