% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.summary.bsmfit}
\alias{print.summary.bsmfit}
\title{Print of the summary information}
\usage{
\method{print}{summary.bsmfit}(x, ...)
}
\arguments{
\item{x}{A BSM fit object}

\item{...}{extra arguments}
}
\description{
.. content for description{} (no empty lines) ..
}
\details{
.. content for details{} ..
}
\author{
Tom Elliott
}

