% Generated by roxygen2 (4.0.2): do not edit by hand
\name{repar}
\alias{repar}
\title{Reparameterize Coefficients}
\usage{
repar(beta, L50, SR, delta = 1)
}
\arguments{
\item{beta}{vector of length 2, \code{c("intercept", "slope")}}

\item{L50}{numeric, length of 50\% retention}

\item{SR}{numeric, selection range}

\item{delta}{numeric, additional Richards' parameter, default = 1}
}
\value{
vector of length 2, \code{c("L50", "SR")} or \code{beta} if L50 and SR specified
}
\description{
Converts the glm coefficients from intercept and slope to the more interpretable L50 and SR.
}
\author{
Tom Elliott
}

