% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{schoolprawns}
\alias{schoolprawns}
\title{Covered codend counts for School Prawns}
\format{A data frame with 342 rows and 12 columns}
\description{
A dataset containing data relating to a covered codend experiment. The species are school prawns,
measured to the nearest mm.
}
\details{
\itemize{
  \item BLOCK. identifying whether the catch was in block 2 or 3
  \item DAY. the day of the haul (1, 2, ..., 10)
  \item SHOT. something?
  \item codend. the number of fish counted in the codend
  \item cover. the number of fish counted in the cover
  \item REP. the number of the replicate trawl
  \item WtSch. the weight of the haul
  \item q1. sampling fraction for the codend
  \item q2. sampling fraction for the cover
  \item total. the total number of prawns counted
  \item Lenclass. the length class (to the nearest mm) of prawns
  \item Haul. id of the haul
}
}
\keyword{datasets}

