% Generated by roxygen2 (4.0.1): do not edit by hand
\name{[.bsmdata}
\alias{[.bsmdata}
\title{Subset bsmData object}
\usage{
\method{[}{bsmdata}(x, i, j)
}
\arguments{
\item{x}{object}

\item{i}{rows}

\item{j}{columns}
}
\value{
subsetted object
}
\description{
Subset a bdmdata object
}
\author{
Tom Elliott
}

