% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.bsmfit}
\alias{summary.bsmfit}
\title{Summary for BSM Fit}
\usage{
\method{summary}{bsmfit}(object, p.values = FALSE, ...)
}
\arguments{
\item{object}{a fitted bsm object}

\item{p.values}{logical, include significance tests for parameters != 0?}

\item{...}{additional parameters}
}
\description{
Summary information for \code{bsmfit} object
}
\details{
Not much ...
}
\author{
Tom Elliott
}

