% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tail.bsmdata}
\alias{tail.bsmdata}
\title{Tail of BSM Data}
\usage{
\method{tail}{bsmdata}(x, ...)
}
\arguments{
\item{x}{a bsmdata object}

\item{...}{extra arguments}
}
\value{
data.frame
}
\description{
Return the last few rows of the data set.
}
\details{
Acts just like tail.data.frame
}
\author{
Tom Elliott
}

