% Generated by roxygen2 (4.0.2): do not edit by hand
\name{testH0}
\alias{testH0}
\title{Generate a Bayesian p-value for a Hypothesis Test}
\usage{
testH0(fit, parameter, H0, alternative = "!=")
}
\arguments{
\item{fit}{a bsmfit object}

\item{parameter}{the name of the parameter to test}

\item{H0}{the null hypothesis value}

\item{alternative}{the alternative hypotheis, one of "!=", "<" or ">"}
}
\value{
numeric
}
\description{
Computes the posterior probability of the null hypothesis being true.
}
\details{
The parameter has a posterior distribution, and this function computes
the posterior probability that the parameter is consistently larger or
smaller than the hypothesised value, which ever is smaller,

min(Pr(theta > H0), Pr(theta < H0))

In the case of a one-sided test, the probability is simply
the posterior probability that the parameter is greater than,
or smaller than, the hypothesised value.
}
\author{
Tom Elliott
}

