% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightmr.R
\name{weightmr}
\alias{weightmr}
\title{Computation of gene ranking weights through Maximum Relevance and Minimum Redundancy (MRMR) method.}
\usage{
weightmr(x, y)
}
\arguments{
\item{x}{Nxp data frame of gene expression values, where, N represents number of genes and p represents samples/time points generated in a case vs. control gene expression study.}

\item{y}{px1 numeric vector with entries 1 and -1 representing sample/subject labels, where 1 and -1 represents the labels of subjects/ samples for case and control conditions respectively.}
}
\value{
A vector of weights computed through MRMR method and a complete ranked gene list.
}
\description{
The function computes weights for features (e.g. genes) using Maximum Relevance and Minimum Redundancy (MRMR) technique and also produce the ranked genelist based on the computed weights.
}
\details{
Computation of gene ranking weights through Maximum Relevance and Minimum Redundancy (MRMR) method.
}
\examples{
x=as.data.frame(matrix(runif(1000), 50))
row.names(x) = paste("Gene", 1:50)
colnames(x) = paste("Samp", 1:20)
y=as.numeric(c(rep(1, 10), rep(-1, 10)))
weightmr(x, y)

}
\references{
Feature selection based on mutual information: criteria of max-dependency, max-relevance and min-redundancy. IEEE Trans. Pattern Anal. Mach. Intell., 27 (8) 1226-1238. DOI: 10.1109/TPAMI.2005.159.
}
\author{
Samarendra Das <samarendra4849 at gmail.com>
}
