% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightsvm.R
\name{weightsvm}
\alias{weightsvm}
\title{Computation of weights for gene selection through SVM algorithm}
\usage{
weightsvm(x, y)
}
\arguments{
\item{x}{Nxp data frame of gene expression values, where, N represents number of genes and p represents samples/time points generated in a case vs. control gene expression study.}

\item{y}{px1 numeric vector with entries 1 and -1 representing sample/subject labels, where 1 and -1 represents the labels of subjects/ samples for case and control condition respectively.}
}
\value{
A vector of weights computed through SVM method and a complete ranked gene list.
}
\description{
computes weights for features (e.g. genes) using Support Vector Machine (SVM) method and also produce the ranked genelist based on the computed weights.
}
\examples{
x=as.data.frame(matrix(runif(1000), 50))
row.names(x) = paste("Gene", 1:50)
colnames(x) = paste("Samp", 1:20)
y=as.numeric(c(rep(1, 10), rep(-1, 10)))
weightsvm(x, y)

}
\references{
Guyon et al. (2002). Gene selection for cancer classification using support vector machines. Mach. Learn., 46, 389-422
}
\author{
Samarendra Das <samarendra4849 at gmail.com>
}
