% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightsvmmrmr.R
\name{weightsvmmrmr}
\alias{weightsvmmrmr}
\title{Computation weights for gene selection through Support Vector Machine  and Maximum Relevance and Minimum Redundancy (MRMR) methods.}
\usage{
weightsvmmrmr(x, y, method, beta)
}
\arguments{
\item{x}{Nxp data frame of gene expression values, where, N represents number of genes and p represents samples/time points generated in a case vs. control gene expression study.}

\item{y}{px1 numeric vector with entries 1 and -1 representing sample/subject labels, where 1 and -1 represents the labels of subjects/ samples for case and control conditions respectively.}

\item{method}{Character variable representing either 'Linear' or 'Quadratic' method for integrating the weights/scores computed through SVM and MRMR methods.}

\item{beta}{scalar representing trade-off between SVM and MRMR weights.}
}
\value{
This function produces a vector weights represents the strength of gene informativeness from gene expression data by integrating scores from SVM and MRMR methods using linear and quadartic techniques of score integration.
}
\description{
The function computes the gene selection weights for genes through a hybrid approach of Support Vector Machine (SVM) and Maximum Relevance and Minimum Redundancy (MRMR) algorithms using the high dimensional gene expression data.
}
\details{
Computation weights for gene selection through SVM-MMRMR method.
Takes the gene expression data matrix (rows as genes and coloumns as samples) and vector of class labels of subjects (1: case and -1: control) as inputs.
}
\examples{
x=as.data.frame(matrix(runif(1000), 50))
row.names(x) = paste("Gene", 1:50)
colnames(x) = paste("Samp", 1:20)
y=as.numeric(c(rep(1, 10), rep(-1, 10)))
weightsvmmrmr(x, y, method="Linear", beta=0.6)

}
\references{
Mundra,P.A. &  Rajapakse, J.C. (2010).SVM-RFE with MRMR filter for gene selection. IEEE Trans. Nanobiosci., 9 (1) 31-37, 10.1109/TNB.2009.2035284.
}
\author{
Samarendra Das <samarendra4849 at gamil.com>
}
