% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dic}
\alias{dic}
\title{Extract DIC Value}
\usage{
dic(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{extra arguments}
}
\value{
varies
}
\description{
Extract the DIC value from an object
}
\details{
Generic Method
}
\author{
tell029
}

