% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generateFormula}
\alias{generateFormula}
\title{Generate a mathematical expression for a parameter}
\usage{
generateFormula(par, expression, data, est, ...)
}
\arguments{
\item{par}{parameter being formulated}

\item{expression}{an expression}

\item{data}{data from fitted object}

\item{est}{parameter estimates for formula}

\item{...}{additional arguments}
}
\value{
a bsmFormula object for a given parameter
}
\description{
.. content for description{} (no empty lines) ..
}
\details{
.. content for details{} ..
}
\author{
Tom Elliott
}

