% Generated by roxygen2 (4.0.2): do not edit by hand
\name{head.bsmdata}
\alias{head.bsmdata}
\title{Head of BSM Data}
\usage{
\method{head}{bsmdata}(x, ...)
}
\arguments{
\item{x}{a bsmdata object}

\item{...}{extra arguments}
}
\value{
data.frame
}
\description{
Return the first few rows of the data set.
}
\details{
Acts just like head.data.frame
}
\author{
Tom Elliott
}

