% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tail.bsmdata}
\alias{tail.bsmdata}
\title{Tail of bsmdata.}
\usage{
\method{tail}{bsmdata}(x, ...)
}
\arguments{
\item{x}{bsmdata object}

\item{...}{extra arguments}
}
\description{
Prints the tail of a bsmdata object.
}
\author{
Tom Elliott
}

